BarrierOption {RQuantLib}	R Documentation

Barrier Option evaluation using Closed-Form solution

Description

This function evaluations an Barrier option on a common stock using a closed-form solution. The option value as well as the common first derivatives ("Greeks") are returned.

Usage

BarrierOption.default(barrType, type, underlying, strike, dividendYield,
                      riskFreeRate, maturity, volatility,
                      barrier, rebate=0.0)

## S3 method for class 'Option':
print
## S3 method for class 'Option':
summary

Arguments

barrType 	 A string with one of the values downin, downout, upin or upout

type 	 A string with one of the values call or put

underlying 	 Current price of the underlying stock

strike 	 Strike price of the option

dividendYield 	 Continuous dividend yield (as a fraction) of the stock

riskFreeRate 	 Risk-free rate

maturity 	 Time to maturity (in fractional years)

volatility 	 Volatility of the underlying stock

barrier 	 Option barrier value

rebate 	 Optional option rebate, defaults to 0.0

Details

A closed-form solution is used to value the Barrier Option. In the case of Barrier options, the calculations are from Haug's "Option pricing formulas" book (McGraw-Hill).

Please see any decent Finance textbook for background reading, and the QuantLib documentation for details on the QuantLib implementation.

Value

An object of class BarrierOption (which inherits from class Option) is returned. It contains a list with the following components:

value 	 Value of option

delta 	 Sensitivity of the option value for a change in the underlying

gamma 	 Sensitivity of the option delta for a change in the underlying

vega 	 Sensitivity of the option value for a change in the underlying's volatility

theta 	 Sensitivity of the option value for a change in t, the remaining time to maturity

rho 	 Sensitivity of the option value for a change in the risk-free interest rate

dividendRho 	 Sensitivity of the option value for a change in the dividend yield

parameters 	 List with parameters with which object was created

.
Note that under the new pricing framework used in QuantLib, binary pricers do not provide analytics for 'Greeks'. This is expected to be addressed in future releases of QuantLib.

Note

The interface might change in future release as QuantLib stabilises its own API.

Author(s)

Dirk Eddelbuettel edd@debian.org for the R interface; the QuantLib Group for QuantLib

References

http://quantlib.org for details on QuantLib.

See Also

AmericanOption,EuropeanOption

Examples

BarrierOption(barrType="downin", type="call", underlying=100,
        strike=100, dividendYield=0.02, riskFreeRate=0.03,
        maturity=0.5, volatility=0.4, barrier=90)

[Package RQuantLib version 0.1.14 Index]