BinaryOptionImpliedVolatility {RQuantLib}	R Documentation

Implied Volatility calculation for Binary Option

Description

The BinaryOptionImpliedVolatility function solves for the (unobservable) implied volatility, given an option price as well as the other required parameters to value an option.

Usage

BinaryOptionImpliedVolatility.default(type, value, underlying, strike,
                dividendYield, riskFreeRate, maturity, volatility,
                cashPayoff=1)

## S3 method for class 'ImpliedVolatility':
print
## S3 method for class 'ImpliedVolatility':
summary

Arguments

type 	 A string with one of the values call, put or straddle

value 	 Value of the option (used only for ImpliedVolatility calculation)

underlying 	 Current price of the underlying stock

strike 	 Strike price of the option

dividendYield 	 Continuous dividend yield (as a fraction) of the stock

riskFreeRate 	 Risk-free rate

maturity 	 Time to maturity (in fractional years)

volatility 	 Initial guess for the volatility of the underlying stock

cashPayoff 	 Binary payout if options is exercised, default is 1

Details

The Finite Differences method is used to value the Binary Option. Implied volatilities are then calculated numerically.

Please see any decent Finance textbook for background reading, and the QuantLib documentation for details on the QuantLib implementation.

Value

The BinaryOptionImpliedVolatility function returns an object of class ImpliedVolatility. It contains a list with the following elements:

impliedVol 	 The volatility implied by the given market prices

parameters 	 List with the option parameters used

Note

The interface might change in future release as QuantLib stabilises its own API.

Author(s)

Dirk Eddelbuettel edd@debian.org for the R interface; the QuantLib Group for QuantLib

References

http://quantlib.org for details on QuantLib.

See Also

EuropeanOption,AmericanOption,BinaryOption

Examples

BinaryOptionImpliedVolatility("call", value=4.50, strike=100, 100, 0.02, 0.03, 0.5, 0.4, 10)

[Package RQuantLib version 0.1.14 Index]