EuropeanOption {RQuantLib}	R Documentation

European Option evaluation using Closed-Form solution

Description

The EuropeanOption function evaluations an European-style option on a common stock using the Black-Scholes-Merton solution. The option value, the common first derivatives ("Greeks") as well as the calling parameters are returned.

Usage

EuropeanOption.default(type, underlying, strike, dividendYield, riskFreeRate, maturity, volatility)

## S3 method for class 'Option':
plot
## S3 method for class 'Option':
print
## S3 method for class 'Option':
summary

Arguments

type 	 A string with one of the values call or put

underlying 	 Current price of the underlying stock

strike 	 Strike price of the option

dividendYield 	 Continuous dividend yield (as a fraction) of the stock

riskFreeRate 	 Risk-free rate

maturity 	 Time to maturity (in fractional years)

volatility 	 Volatility of the underlying stock

Details

The well-known closed-form solution derived by Black, Scholes and Merton is used for valuation. Implied volatilities are calculated numerically.

Please see any decent Finance textbook for background reading, and the QuantLib documentation for details on the QuantLib implementation.

Value

The EuropeanOption function returns an object of class EuropeanOption (which inherits from class Option). It contains a list with the following components:

value 	 Value of option

delta 	 Sensitivity of the option value for a change in the underlying

gamma 	 Sensitivity of the option delta for a change in the underlying

vega 	 Sensitivity of the option value for a change in the underlying's volatility

theta 	 Sensitivity of the option value for a change in t, the remaining time to maturity

rho 	 Sensitivity of the option value for a change in the risk-free interest rate

dividendRho 	 Sensitivity of the option value for a change in the dividend yield

parameters 	 List with parameters with which object was created

Note

The interface might change in future release as QuantLib stabilises its own API.

Author(s)

Dirk Eddelbuettel edd@debian.org for the R interface; the QuantLib Group for QuantLib

References

http://quantlib.org for details on QuantLib.

See Also

EuropeanOptionImpliedVolatility, EuropeanOptionArrays, AmericanOption,BinaryOption

Examples

# simple call with unnamed parameters
EuropeanOption("call", 100, 100, 0.01, 0.03, 0.5, 0.4)
# simple call with some explicit parameters, and slightly increased vol:
EuropeanOption(type="call", underlying=100, strike=100, dividendYield=0.01, 
riskFreeRate=0.03, maturity=0.5, volatility=0.5)

[Package RQuantLib version 0.1.14 Index]