EuropeanOptionImpliedVolatility {RQuantLib}	R Documentation

Implied Volatility calculation for European Option

Description

The EuropeanOptionImpliedVolatility function solves for the (unobservable) implied volatility, given an option price as well as the other required parameters to value an option.

Usage

EuropeanOptionImpliedVolatility.default(type, value, underlying, strike, dividendYield, riskFreeRate, maturity, volatility)

## S3 method for class 'ImpliedVolatility':
print
## S3 method for class 'ImpliedVolatility':
summary

Arguments

type 	 A string with one of the values call or put

value 	 Value of the option (used only for ImpliedVolatility calculation)

underlying 	 Current price of the underlying stock

strike 	 Strike price of the option

dividendYield 	 Continuous dividend yield (as a fraction) of the stock

riskFreeRate 	 Risk-free rate

maturity 	 Time to maturity (in fractional years)

volatility 	 Initial guess for the volatility of the underlying stock

Details

The well-known closed-form solution derived by Black, Scholes and Merton is used for valuation. Implied volatilities are then calculated numerically.

Please see any decent Finance textbook for background reading, and the QuantLib documentation for details on the QuantLib implementation.

Value

The EuropeanOptionImpliedVolatility function returns an object of class ImpliedVolatility. It contains a list with the following elements:

impliedVol 	 The volatility implied by the given market prices

parameters 	 List with the option parameters used

Note

The interface might change in future release as QuantLib stabilises its own API.

Author(s)

Dirk Eddelbuettel edd@debian.org for the R interface; the QuantLib Group for QuantLib

References

http://quantlib.org for details on QuantLib.

See Also

EuropeanOption,AmericanOption,BinaryOption

Examples

EuropeanOptionImpliedVolatility(type="call", value=11.10, underlying=100,
        strike=100, dividendYield=0.01, riskFreeRate=0.03,
        maturity=0.5, volatility=0.4)

[Package RQuantLib version 0.1.14 Index]