ImpliedVolatility {RQuantLib}	R Documentation

Base class for option-price implied volatility evalution

Description

This class forms the basis from which the more specific classes are derived.

Usage

## S3 method for class 'ImpliedVolatility':
print
## S3 method for class 'ImpliedVolatility':
summary

Arguments

{Any option-price implied volatility object derived from this base class}

Details

Please see any decent Finance textbook for background reading, and the QuantLib documentation for details on the QuantLib implementation.

Value

None, but side effects of displaying content.

Note

The interface might change in future release as QuantLib stabilises its own API.

Author(s)

Dirk Eddelbuettel edd@debian.org for the R interface; the QuantLib Group for QuantLib

References

http://quantlib.org for details on QuantLib.

See Also

AmericanOptionImpliedVolatility, EuropeanOptionImpliedVolatility, AmericanOption,EuropeanOption, BinaryOption

Examples

impVol<-EuropeanOptionImpliedVolatility("call", value=11.10, strike=100, volatility=0.4, 100, 0.01, 0.03, 0.5)
print(impVol)
summary(impVol)

[Package RQuantLib version 0.1.14 Index]